feat: add new line support
- move markdown rendered to own component
